fix unwanted polars import
